Refactor: Split VisualCharacter into generic and Character-related components
