+++ src/CMakeLists.txt
@@ -19,6 +19,8 @@ add_library(REGothEngine
   components/CharacterState.cpp
   components/VisualCharacter.hpp
   components/VisualCharacter.cpp
+  components/VisualSkeletalAnimation.hpp
+  components/VisualSkeletalAnimation.cpp
   components/VisualStaticMesh.hpp
   components/VisualStaticMesh.cpp
   components/StartSpot.hpp
@@ -115,3 +117,6 @@ target_link_libraries(REGothWaynetTester REGothEngine samples-common)
 
 add_executable(REGothCharacterMovementTester main_CharacterMovementTest.cpp)
 target_link_libraries(REGothCharacterMovementTester REGothEngine samples-common)
+
+add_executable(REGothMobViewer main_MobViewer.cpp)
+target_link_libraries(REGothMobViewer REGothEngine samples-common)
