+++ src/RTTI/RTTI_TypeIDs.hpp
@@ -5,10 +5,11 @@ namespace REGoth
   // See BsCorePrerequisites.h, TypeID_Core
   enum TypeID_REGoth
   {
-    TID_REGOTH_VisualCharacter        = 60000,
-    TID_REGOTH_Character              = 60001,
-    TID_REGOTH_ScriptBackedBy         = 60002,
-    TID_REGOTH_CharacterKeyboardInput = 60003,
-    TID_REGOTH_CharacterAI            = 60004,
+    TID_REGOTH_VisualCharacter         = 60000,
+    TID_REGOTH_Character               = 60001,
+    TID_REGOTH_ScriptBackedBy          = 60002,
+    TID_REGOTH_CharacterKeyboardInput  = 60003,
+    TID_REGOTH_CharacterAI             = 60004,
+    TID_REGOTH_VisualSkeletalAnimation = 60005,
   };
 }  // namespace REGoth
